gsp/pyproject.toml
[tool.poetry]
name = "gsp"
version = "0.2.0"
description = ""
authors = ["Damian Trafialek <damian.trafialek@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
poethepoet = "^0.26.0"
flake8-pyproject = "^1.2.3"
requests = "^2.31.0"
black = {extras = ["jupyter"], version = "^24.4.2"}
flake8 = "^7.0.0"
mypy = "^1.10.0"
pandas = "^2.2.2"
matplotlib = "^3.8.4"
seaborn = "^0.13.2"
ipykernel = "^6.29.4"
pytest = "^8.2.0"
pytest-cov = "^5.0.0"
numpy = "^1.26.4"
xgboost = "^2.0.3"
lightgbm = "^4.3.0"
optuna = "^3.6.1"
scikit-learn = "^1.4.2"
jupyter = "^1.0.0"
python-dotenv = "^1.0.1"
mongoengine = "^0.28.2"
blinker = "^1.8.2"
injector = "^0.21.0"

[tool.poe.tasks]
scrape = "python -m gsp.scraper.scraper"
lint = "flake8 ."
fmt = "black ."
fmt-check = "black --check ."
typecheck = "mypy ."
convert = "find . -name '*.ipynb' -exec jupyter nbconvert --to python {} \\;"
test = "pytest -v -s ./tests --log-cli-level=INFO"
test-dev = "pytest ./tests -v -s -m dev  --log-cli-level=INFO"
test-cov = "pytest tests -v -s --cov=gsp --log-cli-level=INFO"
pipeline = "python -m gsp.run"

[tool.pytest.ini_options]
markers = [
    "dev: mark a test as a development test.",
]

[tool.flake8]
max-complexity = 10
ignore = "E501,W503,W504,E203,W391"

[tool.black]
line-length = 120

[tool.mypy]
ignore_missing_imports = true
disable_error_code = "import-untyped"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
